changed sagaNum type to Double

+++ src/main/java/com/llacerximo/booklist/persistence/model/BookEntity.java
@@ -17,7 +17,7 @@ public class BookEntity {
     Long id;
     String name;
     Integer pages;
-    Integer sagaNum;
+    Double sagaNum;
     Boolean fave;
     Date publishDate;
     Date startDate;
